==> 1.1 ANÁLISIS CUANTITATIVO Green.ipynb <==
# 1. ANÁLISIS CUANTITATIVO.

## 1.1 Primer examen preliminar del dataset. ¿Qué parámetros hay en el dataset? ¿Cuál es su significado? ¿Existen valores aparentemente incorrectos? 

**1. Importar librerías**

import pandas as pd
import numpy as np

**2. Leer datos GREEN**

### Datos

Tenemos datos de los taxis verdes, de los meses de Marzo y Abril del año 2015.

<table>
    <tr>
        <th>Nombre del archivo</th>
        <th>Tamaño del archivo</th>
        <th>Variables</th>
        </tr>
    <tr>
        <td> green_tripdata_2015-03</td>
        <td> 265 MB </td>
        <td> 20 </td>
    </tr>
    <tr>
        <td> green_tripdata_2015-04</td>
        <td> 256 MB </td>
        <td> 20 </td>
    </tr>

dg3 = pd.read_csv('C:/MBD/datos/green_tripdata_2015-03.csv', usecols=[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20])
dg4 = pd.read_csv('C:/MBD/datos/green_tripdata_2015-04.csv', usecols=[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20])

**3. Head de datos, visualización primeras filas**

dg3.head()

dg4.head()

**5. Descripción columna Ethail Fee y posterior eliminación**

dg3['Ehail_fee'].describe()

dg4['Ehail_fee'].describe()

dg3 = dg3.drop("Ehail_fee",1)

dg4 = dg4.drop("Ehail_fee",1)

**6. Descripción columna Total amount y posterior corrección**

dg3['Total_amount'].describe()

dg3 = dg3.drop(dg3[dg3['Total_amount']<=0].index)

dg4['Total_amount'].describe()

dg4 = dg4.drop(dg4[dg4['Total_amount']<=0].index)

**7. Eliminar datos fuera de rango de latitud y longitud**

dg3 = dg3.drop(dg3[dg3['Pickup_latitude']<40.49].index)
dg3 = dg3.drop(dg3[dg3['Pickup_latitude']>40.92].index)
dg3 = dg3.drop(dg3[dg3['Pickup_longitude']<-74.27].index)
dg3 = dg3.drop(dg3[dg3['Pickup_longitude']>-73.68].index)
dg3 = dg3.drop(dg3[dg3['Dropoff_latitude']<40.49].index)
dg3 = dg3.drop(dg3[dg3['Dropoff_latitude']>40.92].index)
dg3 = dg3.drop(dg3[dg3['Dropoff_longitude']<-74.27].index)
dg3 = dg3.drop(dg3[dg3['Dropoff_longitude']>-73.68].index)

dg3['Pickup_latitude'].describe()

dg3['Pickup_longitude'].describe()

dg3['Dropoff_latitude'].describe()

dg3.head()

dg4 = dg4.drop(dg4[dg4['Pickup_latitude']<40.49].index)
dg4 = dg4.drop(dg4[dg4['Pickup_latitude']>40.92].index)
dg4 = dg4.drop(dg4[dg4['Pickup_longitude']<-74.27].index)
dg4 = dg4.drop(dg4[dg4['Pickup_longitude']>-73.68].index)
dg4 = dg4.drop(dg4[dg4['Dropoff_latitude']<40.49].index)
dg4 = dg4.drop(dg4[dg4['Dropoff_latitude']>40.92].index)
dg4 = dg4.drop(dg4[dg4['Dropoff_longitude']<-74.27].index)
dg4 = dg4.drop(dg4[dg4['Dropoff_longitude']>-73.68].index)

dg4['Pickup_latitude'].describe()

dg4['Pickup_longitude'].describe()

dg4['Dropoff_latitude'].describe()

dg4.head()

**8. Visualización datos correctos + describe de los mismos**

dg3['Passenger_count'].describe()

dg4['Passenger_count'].describe()

dg3['Total_amount'].describe()

dg4['Total_amount'].describe()

dg3 = dg3.drop(dg3[dg3['Total_amount']<=0].index)

dg4 = dg4.drop(dg4[dg4['Total_amount']<=0].index)

dg3['Tip_amount'].describe()

dg4['Tip_amount'].describe()

**9. Escribir csv con limpieza de datos hecha**

dg3.to_csv('C:/MBD/datos/green_tripdata_2015-03_2.csv', index=None , header=True )
dg4.to_csv('C:/MBD/datos/green_tripdata_2015-04_2.csv', index=None , header=True )
