==> 1.1 ANÁLISIS CUANTITATIVO Yellow.ipynb <==
# 1. ANÁLISIS CUANTITATIVO.

## 1.1 Primer examen preliminar del dataset. ¿Qué parámetros hay en el dataset? ¿Cuál es su significado? ¿Existen valores aparentemente incorrectos? 

**1. Importar librerías**

import pandas as pd
import numpy as np

**2. Leer datos YELLOW**

### Datos

Tenemos datos de los taxis verdes, de los meses de Marzo y Abril del año 2015.

<table>
    <tr>
        <th>Nombre del archivo</th>
        <th>Tamaño del archivo</th>
        <th>Variables</th>
        </tr>
    <tr>
        <td> yellow_tripdata_2015-03</td>
        <td> 1,94 GB </td>
        <td> 19 </td>
    </tr>
    <tr>
        <td> yellow_tripdata_2015-04</td>
        <td> 1,90 GB </td>
        <td> 19 </td>
    </tr>

dy3 = pd.read_csv('C:/MBD/datos/yellow_tripdata_2015-03.csv')
dy4 = pd.read_csv('C:/MBD/datos/yellow_tripdata_2015-04.csv')

**3. Head de datos, visualización primeras filas**

dy3.head()

dy4.head()

**4. Describe de columnas relevantes**

dy3['trip_distance'].describe()

dy4['trip_distance'].describe()

**5. Drop de datos con distancia negativa**

dy3 = dy3.drop(dy3[dy3['trip_distance']<0].index)

dy4 = dy4.drop(dy4[dy4['trip_distance']<0].index)

dy3['trip_distance'].describe()

dy4['trip_distance'].describe()

dy3['total_amount'].describe()

dy4['total_amount'].describe()

**6. Drop de datos con Total amount negativo**

dy3 = dy3.drop(dy3[dy3['total_amount']<=0].index)

dy3['total_amount'].describe()

dy4 = dy4.drop(dy4[dy4['total_amount']<=0].index)

dy4['total_amount'].describe()

**7. Selección datos de longitud y latitud**

dy3 = dy3.drop(dy3[dy3['pickup_latitude']<40.49].index)
dy3 = dy3.drop(dy3[dy3['pickup_latitude']>40.92].index)
dy3 = dy3.drop(dy3[dy3['pickup_longitude']<-74.27].index)
dy3 = dy3.drop(dy3[dy3['pickup_longitude']>-73.68].index)
dy3 = dy3.drop(dy3[dy3['dropoff_latitude']<40.49].index)
dy3 = dy3.drop(dy3[dy3['dropoff_latitude']>40.92].index)
dy3 = dy3.drop(dy3[dy3['dropoff_longitude']<-74.27].index)
dy3 = dy3.drop(dy3[dy3['dropoff_longitude']>-73.68].index)

dy4 = dy4.drop(dy4[dy4['pickup_latitude']<40.49].index)
dy4 = dy4.drop(dy4[dy4['pickup_latitude']>40.92].index)
dy4 = dy4.drop(dy4[dy4['pickup_longitude']<-74.27].index)
dy4 = dy4.drop(dy4[dy4['pickup_longitude']>-73.68].index)
dy4 = dy4.drop(dy4[dy4['dropoff_latitude']<40.49].index)
dy4 = dy4.drop(dy4[dy4['dropoff_latitude']>40.92].index)
dy4 = dy4.drop(dy4[dy4['dropoff_longitude']<-74.27].index)
dy4 = dy4.drop(dy4[dy4['dropoff_longitude']>-73.68].index)

dy3['pickup_latitude'].describe()

dy4['pickup_latitude'].describe()

dy3['pickup_longitude'].describe()

dy4['pickup_longitude'].describe()

dy3['dropoff_latitude'].describe()

dy4['dropoff_latitude'].describe()

dy3['passenger_count'].describe()

dy4['passenger_count'].describe()

dy3['tip_amount'].describe()

dy4['tip_amount'].describe()

**8. Eliminar datos con tip amount negativo**

dy3 = dy3.drop(dy3[dy3['tip_amount']<0].index)

dy4 = dy4.drop(dy4[dy4['tip_amount']<0].index)

dy3['store_and_fwd_flag'].value_counts()

dy4['store_and_fwd_flag'].value_counts()

dy3['payment_type'].value_counts()

dy4['payment_type'].value_counts()

dy3['fare_amount'].describe()

dy4['fare_amount'].describe()

dy3.shape

dy4.shape

**9. Escribir csv con limpieza de datos hecha**

dy3.to_csv('C:/MBD/datos/yellow_tripdata_2015-03_2.csv', index=None , header=True )
dy4.to_csv('C:/MBD/datos/yellow_tripdata_2015-03_2.csv', index=None , header=True )
