==> 2. Cualitativo Green.ipynb <==
# 2. ANÁLISIS CUALITATIVO.

## 2.1 ¿Cuál es el trayecto en el que la relación precio/km es más alta? ¿Cuál es el trayecto en el que la relación tiempo/km es más alta? ¿Cuál es el trayecto en el que la relación precio/tiempo es más alta?



## 2.2 ¿Cuál es el trayecto en el que la relación precio/km es más baja? ¿Cuál es el trayecto en el que la relación tiempo/km es más baja? ¿Cuál es el trayecto en el que la relación precio/tiempo es más baja?

**1. Importar librerías**

import pandas as pd
import numpy as np
from datetime import datetime,date
import time

**2. Leer datos GREEN**

dg3 = pd.read_csv('C:/Master/green_tripdata_2015-03_2.csv')

**3. Head de datos, visualización primeras filas**

dg3.head()

**4. Calculamos la carrera más cara y más barata**
#### Nosotros contamos la propina dentro del precio total del servicio

dg3['Expensive'] = dg3.apply(
    lambda s: 0 if s['Trip_distance']==0 else round((s['Total_amount']+s['Tip_amount'])/s['Trip_distance'],2), axis=1)

dg3.head()

**5. ¿Cuál es el trayecto en el que la relación precio/km es más alta?**

print(np.max(dg3['Expensive']))

expensive = dg3[dg3['Expensive'] ==  np.max(dg3['Expensive'])]

expensive

print(expensive['Total_amount']+expensive['Tip_amount'])
print(expensive['Trip_distance'])

**6. ¿Cuál es el trayecto en el que la relación precio/km es más baja?**

print(np.min(dg3['Expensive']))

cheaper = dg3[dg3['Expensive'] ==  np.min(dg3['Expensive'])]

cheaper

print(cheaper['Trip_distance'])
print(cheaper['Total_amount']+cheaper['Tip_amount'])

 **7. Calculamos las relación tiempo/km más alta y más baja**

dg3.dtypes

Necesitamos restar el tiempo de bajada de bandera con recogida y pasarlo a segundos y mas tarde a minutos, para los calculos que necesitamos realizar

dg3['Time_carreer_seconds'] = dg3.apply(
    lambda s:
    time.mktime((datetime.strptime(s['Lpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S")).timetuple())
    -time.mktime((datetime.strptime(s['lpep_pickup_datetime'],"%Y-%m-%d %H:%M:%S")).timetuple())
    ,axis=1)

dg3.head()

dg3['Time_carrer_minutes'] = dg3.apply(
    lambda s:
    (s['Time_carreer_seconds'] % 3600) // 60
    ,axis=1)

dg3.head()

Ahora podemos realizar el cálculo de minutos respecto a los km o millas recorridos

dg3['Time_long_vs_km'] = dg3.apply(
    lambda s: round(s['Time_carreer_seconds']/s['Trip_distance'],2), axis=1)

expensive = dg3[dg3['Time_long_vs_km'] ==  np.max(dg3['Time_long_vs_km'])]

 **8. ¿Cuál es el trayecto en el que la relación tiempo/km es más alta?**

expensive

print(expensive['Trip_distance'])
print(expensive['Time_carrer_minutes'])

 **9. ¿Cuál es el trayecto en el que la relación tiempo/km es más baja?**

cheaper = dg3[dg3['Time_long_vs_km'] ==  np.min(dg3['Time_long_vs_km'])]

cheaper

print(cheaper['Trip_distance'])
print(cheaper['Time_carrer_minutes'])

 **10. ¿Cuál es el trayecto en el que la relación precio/tiempo es más alta?**

dg3['Time_long_vs_price'] = dg3.apply(
    lambda s: round(s['Time_carreer_seconds']/s['Total_amount'],2), axis=1)

expensive = dg3[dg3['Time_long_vs_price'] ==  np.max(dg3['Time_long_vs_price'])]

expensive

print(expensive['Total_amount'])
print(expensive['Time_carrer_minutes'])

 **11. ¿Cuál es el trayecto en el que la relación precio/tiempo es más baja?**

cheaper = dg3[dg3['Time_long_vs_price'] ==  np.min(dg3['Time_long_vs_price'])]

cheaper

print(cheaper['Total_amount'])
print(cheaper['Time_carrer_minutes'])

## 2.3 Muestra la evolución del tiempo medio de trayecto a lo largo del día. Muestra la evolución de la distancia media de trayecto a lo largo del día.

**1. Creamos variables relevantes para el caso de estudio**

dg3['Week_day'] = dg3.apply(
    lambda s:
    datetime.weekday(datetime.strptime(s['Lpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S"))+1
    ,axis=1)

dg3.head()

dg3['Hour_day'] = dg3.apply(
    lambda s:
    datetime.strptime(s['Lpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S").hour
    ,axis=1)

dg3.head()

def df(d,dia,hora):

    return d[(d['Week_day']==dia) & (d['Hour_day']==hora)]

### Realizamos una estimacion de la media de tiempo y recorrido por dia de la semana y hora.

dia=1
hora=0

while dia<8:
    while hora<24:
        data=(df(dg3,dia,hora))
        print('Dia: ',dia, 'Hora: ',hora,'Tiempo: ',round(sum(data['Time_carrer_minutes'])/len(data),2),'Recorrido: ',round(sum(data['Trip_distance'])/len(data),2))
        hora=hora+1
    dia=dia+1
    hora=0

#sns.catplot(x="Week_day", y="Hour_day", col=avg, data=data)    

# Intentamos encontrar los servicios del punto pelicula Die Hard III
Hemos buscadolas cordenadas de los cuadrantes corresponientes a los puntos origen y destino, los hemos pasado a decimales con aproximación a 14.

Obtenidas de:

https://www.coordenadas-gps.com/convertidor-de-coordenadas-gps

dg3.head()

dg=dg3[(dg3['Pickup_latitude']<=40.779279706100030) & (dg3['Pickup_latitude']>=40.778470146623080)]
dg=dg3[(dg3['Pickup_longitude']>=-73.982385378544050) & (dg3['Pickup_longitude']<=-73.981940513982220)]
dg=dg[(dg['Dropoff_latitude']<=40.707792443681770) & (dg['Dropoff_latitude']>=40.707101832301877)]
dg=dg[(dg['Dropoff_longitude']>=-74.012211704235480) & (dg['Dropoff_longitude']<=-74.011853565194060)]

len(dg)


dg.head()

carrera=dg[(dg['Time_carrer_minutes']<=30.0)]

carrera

carrera['Time_carrer_minutes'].describe()

if len(dg) != 0:
    print('Exito: ',len(carrera),'Total: ',len(dg),'%Exito: ',round(len(carrera)/len(dg),2))

dg.to_csv('C:/master/green_diehardIII_2015-03_2.csv', index=None , header=True )
