==> 2. Cualitativo Yellow.ipynb <==
# 2. ANÁLISIS CUALITATIVO.

## 2.1 ¿Cuál es el trayecto en el que la relación precio/km es más alta? ¿Cuál es el trayecto en el que la relación tiempo/km es más alta? ¿Cuál es el trayecto en el que la relación precio/tiempo es más alta?



## 2.2 ¿Cuál es el trayecto en el que la relación precio/km es más baja? ¿Cuál es el trayecto en el que la relación tiempo/km es más baja? ¿Cuál es el trayecto en el que la relación precio/tiempo es más baja?

**1. Importar librerías**

import pandas as pd
import numpy as np
from datetime import datetime,date
import time

**2. Leer datos Yellow**

dy3 = pd.read_csv('C:/master/yellow_tripdata_2015-03_2.csv')

**3. Head de datos, visualización primeras filas**

dy3.head()

**4. Calculamos la carrera más cara y más barata**

Calculamos segun el precio de la carrera y la propina dada cuales van a ser los viajes mas caros o mas baratos teniendo presente la relación establecida en el enunciado

dy3['Expensive'] = dy3.apply(
    lambda s: 0 if s['trip_distance']==0 else round((s['total_amount']+s['tip_amount'])/s['trip_distance'],2), axis=1)

dy3.head()

**5. ¿Cuál es el trayecto en el que la relación precio/km es más alta?**

print(np.max(dy3['Expensive']))

expensive = dy3[dy3['Expensive'] ==  np.max(dy3['Expensive'])]

expensive

print(expensive['tip_amount']+expensive['total_amount'])
print(expensive['trip_distance'])

**6. ¿Cuál es el trayecto en el que la relación precio/km es más baja?**

dy3 = dy3.drop(dy3[dy3['total_amount']<=1].index)

print(np.min(dy3['Expensive']))

cheaper = dy3[dy3['Expensive'] ==  np.min(dy3['Expensive'])]

cheaper.head()

print(cheaper['tip_amount']+cheaper['total_amount'])
print(cheaper['trip_distance'])

 **7. Calculamos las relación tiempo/km más alta y más baja**

dy3.dtypes

dy3['Time_carreer_seconds'] = dy3.apply(
    lambda s:
    time.mktime((datetime.strptime(s['tpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S")).timetuple())
    -time.mktime((datetime.strptime(s['tpep_pickup_datetime'],"%Y-%m-%d %H:%M:%S")).timetuple())
    ,axis=1)

dy3.head()

dy3['Time_carrer_minutes'] = dy3.apply(
    lambda s:
    (s['Time_carreer_seconds'] % 3600) // 60
    ,axis=1)

dy3.head()

dy3['Time_long_vs_km'] = dy3.apply(
    lambda s: round(s['Time_carreer_seconds']/s['trip_distance'],2), axis=1)

expensive = dy3[dy3['Time_long_vs_km'] ==  np.max(dy3['Time_long_vs_km'])]

 **8. ¿Cuál es el trayecto en el que la relación tiempo/km es más alta?**

expensive

print(expensive['trip_distance'])
print(expensive['Time_carrer_minutes'])

 **9. ¿Cuál es el trayecto en el que la relación tiempo/km es más baja?**

cheaper = dy3[dy3['Time_long_vs_km'] ==  np.min(dy3['Time_long_vs_km'])]

cheaper

print(cheaper['trip_distance'])
print(cheaper['Time_carrer_minutes'])

 **10. ¿Cuál es el trayecto en el que la relación precio/tiempo es más alta?**

dy3['Time_long_vs_price'] = dy3.apply(
    lambda s: round(s['Time_carreer_seconds']/s['total_amount'],2), axis=1)

expensive = dy3[dy3['Time_long_vs_price'] ==  np.max(dy3['Time_long_vs_price'])]

expensive

print(expensive['total_amount']+expensive['tip_amount'])
print(expensive['Time_carrer_minutes'])

 **11. ¿Cuál es el trayecto en el que la relación precio/tiempo es más alta?**

cheaper = dy3[dy3['Time_long_vs_price'] ==  np.min(dy3['Time_long_vs_price'])]

cheaper

print(cheaper['total_amount']+cheaper['tip_amount'])
print(cheaper['Time_carrer_minutes'])

## 2.3 Muestra la evolución del tiempo medio de trayecto a lo largo del día. Muestra la evolución de la distancia media de trayecto a lo largo del día.

**1. Creamos variables relevantes para el caso de estudio**

dy3['Week_day'] = dy3.apply(
    lambda s:
    datetime.weekday(datetime.strptime(s['tpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S"))+1
    ,axis=1)

dy3.head()

dy3['Hour_day'] = dy3.apply(
    lambda s:
    datetime.strptime(s['tpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S").hour
    ,axis=1)

dy3.head()

def df(d,dia,hora):

    return d[(d['Week_day']==dia) & (d['Hour_day']==hora)]

dia=1
hora=0

while dia<8:
    while hora<24:
        data=(df(dy3,dia,hora))
        print('Dia: ',dia, 'Hora: ',hora,'Tiempo: ',round(sum(data['Time_carrer_minutes'])/len(data),2),'Recorrido: ',round(sum(data['trip_distance'])/len(data),2))
        hora=hora+1
    dia=dia+1
    hora=0

 

Observamos que los tiempos son altos para el poco recorrido que se hace, en comparación con los datos de taxis Green

# Intentamos encontrar los servicios del punto pelicula Die Hard III
Hemos buscadolas cordenadas de los cuadrantes corresponientes a los puntos origen y destino, los hemos pasado a decimales con aproximación a 14.

Obtenidas de:

https://www.coordenadas-gps.com/convertidor-de-coordenadas-gps

dy3.head()

dy=dy3[(dy3['pickup_latitude']<=40.779279706100030) & (dy3['pickup_latitude']>=40.778470146623080)]

dy=dy3[(dy3['pickup_longitude']>=-73.982385378544050) & (dy3['pickup_longitude']<=-73.981940513982220)]

dy=dy[(dy['dropoff_latitude']<=40.707792443681770) & (dy['dropoff_latitude']>=40.707101832301877)]

dy=dy[(dy['dropoff_longitude']>=-74.012211704235480) & (dy['dropoff_longitude']<=-74.011853565194060)]

len(dy)

dy.head()

carrera=dy[(dy['Time_carrer_minutes']<=30.0)]

carrera

carrera['Time_carrer_minutes'].describe()

print('Exito: ',len(carrera),'Total: ',len(dy),'%Exito: ',round(len(carrera)/len(dy),2))

dy.to_csv('C:/master/yellow_diehardIII_2015-03_2.csv', index=None , header=True )
