==> 3.2 Análisis predictivo.ipynb <==
# 3. ANÁLISIS PREDICTIVO.

## 3.2 ¿Cuál es la mejor hora del día para ir al aeropuerto?

**1. Importar librerías**

import pandas as pd
import numpy as np

**2. Configuración para trabajar en el entorno de collaboratory**

!pip install -U -q PyDrive

from pydrive.auth import GoogleAuth
from pydrive.drive import GoogleDrive
from google.colab import auth
from oauth2client.client import GoogleCredentials

auth.authenticate_user()
gauth = GoogleAuth()
gauth.credentials = GoogleCredentials.get_application_default()
drive = GoogleDrive(gauth)

from google.colab import files

csv_import = drive.CreateFile({'id':'1TlBINWeLJhWvVMqkeDW-jflVKEH3zv3A'})

 csv_import.GetContentFile('yellow_tripdata_2015-03.csv') 

**3. Leer datos**

dy3 = pd.read_csv('yellow_tripdata_2015-03.csv', nrows = 1000000)

**4. Visualización de los datos y shape de los mismos**

dy3.shape

dy3.head()

**5. Limpieza de datos**

dy3 = dy3.drop(dy3[dy3['total_amount']<=0].index)
dy3 = dy3.drop(dy3[dy3['total_amount']<=1].index)
dy3 = dy3.drop(dy3[dy3['total_amount']>=1000].index)

dy3 = dy3.drop(dy3[dy3['pickup_latitude']<40.49].index)
dy3 = dy3.drop(dy3[dy3['pickup_latitude']>40.92].index)
dy3 = dy3.drop(dy3[dy3['pickup_longitude']<-74.27].index)
dy3 = dy3.drop(dy3[dy3['pickup_longitude']>-73.68].index)
dy3 = dy3.drop(dy3[dy3['dropoff_latitude']<40.49].index)
dy3 = dy3.drop(dy3[dy3['dropoff_latitude']>40.92].index)
dy3 = dy3.drop(dy3[dy3['dropoff_longitude']<-74.27].index)
dy3 = dy3.drop(dy3[dy3['dropoff_longitude']>-73.68].index)
dy3 = dy3.drop(dy3[dy3['trip_distance']<0.1].index)

**6. Importar librerías para cambiar formato de fechas**

from datetime import datetime,date
import time

**7. Crear dos nuevas variables, duraciñon del trayecto en segundos y minutos**

dy3['Time_carreer_seconds'] = dy3.apply(
    lambda s:
    time.mktime((datetime.strptime(s['tpep_dropoff_datetime'],"%Y-%m-%d %H:%M:%S")).timetuple())
    -time.mktime((datetime.strptime(s['tpep_pickup_datetime'],"%Y-%m-%d %H:%M:%S")).timetuple())
    ,axis=1)

dy3['Duración_del_trayecto_min'] = dy3.apply(
    lambda s:
    (s['Time_carreer_seconds'] % 3600) // 60
    ,axis=1)

**8. A  partir de las columnas de fechas, creamos variables que creemos que aportarán valor a nuestro modelo**

def add_datetime_features(dy3):
    
    year = lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).year
    dy3['year'] = dy3['tpep_pickup_datetime'].map(year)
    print('1/7')
    
    hour = lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).hour
    dy3['hour'] = dy3['tpep_pickup_datetime'].map(hour)
    print('2/7')

    day_of_week = lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).weekday()
    dy3['day_of_week'] = dy3['tpep_pickup_datetime'].map(day_of_week)
    print('3/7')

    month = lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).month
    dy3['month'] = dy3['tpep_pickup_datetime'].map(month)
    print('4/7')

    week_number = lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).strftime('%V')
    dy3['week_number'] = dy3['tpep_pickup_datetime'].map(week_number)
    print('5/7')
    
    seasons = [0,0,1,1,1,2,2,2,3,3,3,0] #dec - feb is winter, then spring, summer, fall etc
    season = lambda x: seasons[(datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).month-1)]
    dy3['season'] = dy3['tpep_pickup_datetime'].map(season)
    print('6/7')
    
    late_night_hours = [1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1]
    late_night = lambda x: late_night_hours[(datetime.strptime(x, "%Y-%m-%d %H:%M:%S" ).hour)]
    dy3['late_night'] = dy3['tpep_pickup_datetime'].map(late_night)
    print('7/7. All done!')
    
add_datetime_features(dy3)

**9. Creamos variable velocidad en millas / min**

dy3['velocidad'] = (dy3['trip_distance']/dy3['Duración_del_trayecto_min'])

dy3.head()

**10. Creamos la distancia esférica para el posterior cálculo del punto 11.**

def sphere_dist(pickup_lat, pickup_lon, dropoff_lat, dropoff_lon):
   
    R_earth = 6371
    pickup_lat, pickup_lon, dropoff_lat, dropoff_lon = map(np.radians,
                                                             [pickup_lat, pickup_lon, 
                                                              dropoff_lat, dropoff_lon])
    dlat = dropoff_lat - pickup_lat
    dlon = dropoff_lon - pickup_lon
    
    a = np.sin(dlat/2.0)**2 + np.cos(pickup_lat) * np.cos(dropoff_lat) * np.sin(dlon/2.0)**2
    
    return 2 * R_earth * np.arcsin(np.sqrt(a))

**11. Cálculo de las distancias a los aeropuertos**

def add_airport_dist(dataset):
    """
    JFK: John F. Kennedy International Airport
    EWR: Newark Liberty International Airport
    LGA: LaGuardia Airport
    """
    jfk_coord = (40.639722, -73.778889)
    ewr_coord = (40.6925, -74.168611)
    lga_coord = (40.77725, -73.872611)
    
    pickup_lat = dataset['pickup_latitude']
    dropoff_lat = dataset['dropoff_latitude']
    pickup_lon = dataset['pickup_longitude']
    dropoff_lon = dataset['dropoff_longitude']
    
    pickup_jfk = sphere_dist(pickup_lat, pickup_lon, jfk_coord[0], jfk_coord[1]) 
    dropoff_jfk = sphere_dist(jfk_coord[0], jfk_coord[1], dropoff_lat, dropoff_lon) 
    pickup_ewr = sphere_dist(pickup_lat, pickup_lon, ewr_coord[0], ewr_coord[1])
    dropoff_ewr = sphere_dist(ewr_coord[0], ewr_coord[1], dropoff_lat, dropoff_lon) 
    pickup_lga = sphere_dist(pickup_lat, pickup_lon, lga_coord[0], lga_coord[1]) 
    dropoff_lga = sphere_dist(lga_coord[0], lga_coord[1], dropoff_lat, dropoff_lon) 
    
    dataset['JFKDist'] = pd.concat([pickup_jfk, dropoff_jfk], axis=1).min(axis=1)
    dataset['EWRDist'] = pd.concat([pickup_ewr, dropoff_ewr], axis=1).min(axis=1)
    dataset['LGADist'] = pd.concat([pickup_lga, dropoff_lga], axis=1).min(axis=1)
    
    return dataset

**12. Añadimos el dataset creado al dataset existente**

dy3 = add_airport_dist(dy3)

**13. Visualización de los datos**

dy3.head()

**14. Creación variable tiempo a aeropuerto**

dy3['Tiempo_aeropuerto_JFK'] = dy3['JFKDist'] / dy3['velocidad']
dy3['Tiempo_aeropuerto_EWR'] = dy3['EWRDist'] / dy3['velocidad']
dy3['Tiempo_aeropuerto_LGA'] = dy3['LGADist'] / dy3['velocidad']

**15. Visualización datos**

dy3.head()

**16. Describe de variable ejemplo tiempo aeropuerto a JFK**

dy3['Tiempo_aeropuerto_JFK'].describe()

**17. Análisis quantiles de variables creadas**

quantile_Tiempo_aeropuerto_JFK = dy3.Tiempo_aeropuerto_JFK.quantile(np.round(np.arange(0.00, 1.01, 0.01), 2))
qValues = np.round(np.arange(0.01, 1.01, 0.01), 2)
for i in qValues:
    print("{} Valor del percentil tiempo a JFK: {}min".format((int(i*100)), quantile_Tiempo_aeropuerto_JFK[i]))

quantile_Tiempo_aeropuerto_EWR = dy3.Tiempo_aeropuerto_EWR.quantile(np.round(np.arange(0.00, 1.01, 0.01), 2))
qValues = np.round(np.arange(0.01, 1.01, 0.01), 2)
for i in qValues:
    print("{} Valor del percentil tiempo a EWR: {}min".format((int(i*100)), quantile_Tiempo_aeropuerto_EWR[i]))

quantile_Tiempo_aeropuerto_LGA = dy3.Tiempo_aeropuerto_LGA.quantile(np.round(np.arange(0.00, 1.01, 0.01), 2))
qValues = np.round(np.arange(0.01, 1.01, 0.01), 2)
for i in qValues:
    print("{} Valor del percentil tiempo a LGA: {}min".format((int(i*100)), quantile_Tiempo_aeropuerto_LGA[i]))

**18. Limpieza de datos erroneos bajo criterio porpio**

dy3['Tiempo_aeropuerto_JFK'].describe()

dy3['Tiempo_aeropuerto_EWR'].describe()

dy3['Tiempo_aeropuerto_LGA'].describe()

**19. Limpiamos los datos en base a la media, así en el cálculo posterior se obtiene la mejor hora para el tiempo medio que se tarda a cada aeropuerto.**

dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_JFK']<=0].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_JFK']<=123].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_JFK']>124].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_EWR']<=0].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_EWR']<=98].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_EWR']>100].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_LGA']<=0].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_LGA']<=53].index)
dy3 = dy3.drop(dy3[dy3['Tiempo_aeropuerto_LGA']>54].index)

best_hour_JFK = dy3[dy3['Tiempo_aeropuerto_JFK'] ==  np.min(dy3['Tiempo_aeropuerto_JFK'])]
best_hour_EWR = dy3[dy3['Tiempo_aeropuerto_EWR'] ==  np.min(dy3['Tiempo_aeropuerto_EWR'])]
best_hour_LGA = dy3[dy3['Tiempo_aeropuerto_LGA'] ==  np.min(dy3['Tiempo_aeropuerto_LGA'])]

**20. Resultados de mejor hora para viajar al aeropuerto**

print(best_hour_JFK['hour'])
print(best_hour_JFK['Tiempo_aeropuerto_JFK'])
print(best_hour_JFK['trip_distance'])
print(best_hour_EWR['hour'])
print(best_hour_EWR['Tiempo_aeropuerto_EWR'])
print(best_hour_EWR['trip_distance'])
print(best_hour_LGA['hour'])
print(best_hour_LGA['Tiempo_aeropuerto_LGA'])
print(best_hour_LGA['trip_distance'])
